Lógica del logout (borrar token y cookie), cambio de estado de las mesas asignadas, Preguntas frecuentes, Pedidos en la carta

+++ src/app/core/components/header/header.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core'
 import { MatDialog } from '@angular/material/dialog'
 import { AuthService } from '@pa/auth/services'
+import { MesasService } from '@pa/mesas/services'
 import { CookieService } from 'ngx-cookie-service'
 import { QrScannerComponent } from 'src/app/shared/components/qr-scanner/qr-scanner.component'
 
@@ -13,14 +14,23 @@ export class HeaderComponent {
   constructor(
     public _authService: AuthService,
     public dialog: MatDialog,
-    private _cookieService: CookieService
+    private _cookieService: CookieService,
+    private _mesaService: MesasService
   ) {}
 
   logout() {
     this._authService.logout().subscribe({
       next: () => {
-        this._cookieService.delete('ClienteMesa', '/')
-        localStorage.removeItem('token')
+        if (this._cookieService.check('ClienteMesa')) {
+          const cookieValue = this._cookieService.get('ClienteMesa')
+          const idMesa = Number(cookieValue.split(':')[1])
+          this._mesaService.habilitarMesa(idMesa).subscribe({
+            next: () => {
+              this._cookieService.delete('ClienteMesa', '/')
+            }
+          })
+        }
+        this._authService.borrarToken()
         window.location.href = '/'
       },
       error: (err) => {
@@ -33,7 +43,7 @@ export class HeaderComponent {
     const dialogRef = this.dialog.open(QrScannerComponent, {})
     dialogRef.afterClosed().subscribe((resultado) => {
       if (resultado) {
-        window.location.href = '/'
+        window.location.href = '/carta'
       }
     })
   }
